endpoints - change directories so we dont have to tell spring every time - justin/Joe/Wheeler
