+++ src/main/java/com/springboothackathon/SpringBootApplication.java
@@ -1,20 +1,18 @@
-package springboothackathon;
+package com.springboothackathon;
 
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
-import springboothackathon.controllers.HomeController;
 
 import java.util.Arrays;
 
-@SpringBootApplication(scanBasePackages = {"springboothackathon.controllers", "springboothackathon.services", "springboothackathon.models"})
-@ComponentScan(basePackageClasses = HomeController.class)
-public class Application {
+@org.springframework.boot.autoconfigure.SpringBootApplication(scanBasePackages = "com.springboothackathon")
+@ComponentScan
+public class SpringBootApplication {
     public static void main(String[] args) {
-        SpringApplication.run(Application.class, args);
+        SpringApplication.run(SpringBootApplication.class, args);
     }
 
     @Bean
