+++ src/main/java/com/springboothackathon/controllers/HomeController.java
@@ -1,4 +1,4 @@
-package springboothackathon.controllers;
+package com.springboothackathon.controllers;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
