+++ src/main/java/com/springboothackathon/models/Movie.java
@@ -1,4 +1,4 @@
-package springboothackathon.models;
+package com.springboothackathon.models;
 
 public class Movie {
     private int id;
