+++ src/main/java/com/springboothackathon/services/TheaterService.java
@@ -1,0 +1,12 @@
+package com.springboothackathon.services;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.stereotype.Component;
+import com.springboothackathon.models.Theater;
+
+import java.util.ArrayList;
+
+@Component
+public interface TheaterService {
+    ArrayList<Theater> getAllTheaters();
+}
