--- src/main/java/springboothackathon/services/TheaterServiceImpl.java
@@ -1,12 +1,0 @@
-package springboothackathon.services;
-
-import springboothackathon.models.Theater;
-
-import java.util.List;
-
-public class TheaterServiceImpl implements TheaterService {
-    @Override
-    public List<Theater> getAllTheaters() {
-        return null;
-    }
-}
