+++ src/test/java/com/springboothackathon/controllers/TheaterControllerTests.java
@@ -1,15 +1,13 @@
-package springboothackathon.controllers;
+package com.springboothackathon.controllers;
 
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mock;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
+import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.web.servlet.MockMvc;
-import springboothackathon.models.Theater;
-import springboothackathon.services.TheaterService;
+import com.springboothackathon.models.Theater;
+import com.springboothackathon.services.TheaterService;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -17,17 +15,14 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@SpringBootTest
 @AutoConfigureMockMvc
-@ExtendWith(SpringExtension.class)
+@SpringBootTest
 public class TheaterControllerTests {
     @Autowired
     private MockMvc mockMvc;
 
-    @Mock
+    @MockBean
     private TheaterService service;
 
     @Test
@@ -42,9 +37,7 @@ public class TheaterControllerTests {
         List<Theater> theaters = new ArrayList<Theater>();
         theaters.add(theater);
 
-        when(service.getAllTheaters()).thenReturn(theaters);
-
-        mockMvc.perform(get("/theaters")).andDo(print());
+        mockMvc.perform(get("/theaters"));
 
         verify(service, times(1)).getAllTheaters();
     }
@@ -58,7 +51,7 @@ public class TheaterControllerTests {
         theater.setLatitude(22.3f);
         theater.setLongitude(22.3f);
 
-        List<Theater> theaters = new ArrayList<Theater>();
+        ArrayList<Theater> theaters = new ArrayList<Theater>();
         theaters.add(theater);
 
         when(service.getAllTheaters()).thenReturn(theaters);
