--- src/test/java/springboothackathon/controllers/HomeControllerTest.java
@@ -1,50 +1,0 @@
-package springboothackathon.controllers;
-
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.client.TestRestTemplate;
-import org.springframework.boot.web.server.LocalServerPort;
-import org.springframework.test.context.junit4.SpringRunner;
-import org.springframework.test.web.servlet.MockMvc;
-
-import static org.junit.jupiter.api.Assertions.*;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-@SpringBootTest
-@AutoConfigureMockMvc //tells spring boot not to load the entire webserver and context, but to mock the server
-class HomeControllerTest {
-
-    @Autowired
-    private MockMvc mockMvc;
-
-    @Test
-    void callsHome() throws Exception {
-        mockMvc.perform(get("/home"))
-                .andDo(print())
-                .andExpect(status().isOk());
-
-    }
-}
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class HomeControllerWithMockedWebServer{
-
-    @LocalServerPort
-    private int port;
-
-    @Autowired
-    private TestRestTemplate restTemplate;
-
-    @Test
-    void callsHome(){
-
-        String result = restTemplate.getForObject("http://localhost:" + port + "/home", String.class);
-
-        assertEquals(result, "Greetings from Spring Boot!");
-    }
-}
